clean up tests and fix error handling/timeouts

+++ delete_post.go
@@ -10,7 +10,7 @@ import (
 
 func (cfg *apiConfig) deletePost(w http.ResponseWriter, r *http.Request) {
 	postIDStr := chi.URLParam(r, "postID")
-	var postID int
+	var postID int32
 	_, err := fmt.Sscanf(postIDStr, "%d", &postID)
 	if err != nil {
 		log.Printf("error parsing postID: %v", err)
@@ -20,7 +20,7 @@ func (cfg *apiConfig) deletePost(w http.ResponseWriter, r *http.Request) {
 
 	log.Printf("deleting post: %v", postID)
 
-	err2 := cfg.db.DeletePost(r.Context(), int32(postID))
+	err2 := cfg.db.DeletePost(r.Context(), postID)
 	if err2 != nil {
 		log.Printf("error deleting item from items table: %v", err2)
 		w.WriteHeader(http.StatusInternalServerError)
